+++ plugins/com.google.cloud.tools.eclipse.appengine.deploy/src/com/google/cloud/tools/eclipse/appengine/deploy/messages.properties
@@ -3,8 +3,8 @@ cannot.parse.appengine.xml=Cannot parse appengine-web.xml
 cleanup.deploy.job.name=Delete old deploy work directories
 cleanup.deploy.job.error=Cannot delete old deploy work directories
 task.name.deploy.project=Deploying project
-task.name.stage.project=Staging exploded WAR
-task.name.publish.war=Publish exploded WAR
+task.name.stage.project=Staging deploy artifact
+task.name.publish.war=Preparing deploy artifact
 project.id.missing=App Engine project ID is not set, deploy failed.
 deploy.standard.runnable.name=Deploying to App Engine Standard
 deploy.job.stagingdir.create.failed=Cannot create staging directory.
