+++ plugins/com.google.cloud.tools.eclipse.appengine.deploy/src/com/google/cloud/tools/eclipse/appengine/deploy/standard/StandardDeployJob.java
@@ -26,6 +26,8 @@ import com.google.cloud.tools.appengine.cloudsdk.process.StringBuilderProcessOut
 import com.google.cloud.tools.eclipse.appengine.deploy.AppEngineDeployOutput;
 import com.google.cloud.tools.eclipse.appengine.deploy.AppEngineProjectDeployer;
 import com.google.cloud.tools.eclipse.appengine.deploy.Messages;
+import com.google.cloud.tools.eclipse.appengine.deploy.DeployStaging;
+import com.google.cloud.tools.eclipse.appengine.deploy.WarPublisher;
 import com.google.cloud.tools.eclipse.login.CredentialHelper;
 import com.google.cloud.tools.eclipse.sdk.CollectingLineListener;
 import com.google.cloud.tools.eclipse.ui.util.WorkbenchUtil;
@@ -167,7 +169,7 @@ public class StandardDeployJob extends WorkspaceJob {
     try {
       getJobManager().beginRule(project, progress);
       WarPublisher.publishExploded(project, explodedWarDirectory, progress.newChild(40));
-      new StandardProjectStaging().stage(explodedWarDirectory, stagingDirectory,
+      new DeployStaging().stage(explodedWarDirectory, stagingDirectory,
           cloudSdk, progress.newChild(60));
       return stagingExitListener.getExitStatus();
     } catch (CoreException | IllegalArgumentException | OperationCanceledException ex) {
